requirements.txt
torch==2.6.0
torchvision==0.21.0
opencv-python==4.11.0.86
pillow==9.5.0
requests==2.31.0